+++ engines/tahi_standard_tasks/client/app/components/paper-reviewer-task.js
@@ -1,2 +1,7 @@
 import TaskComponent from 'tahi/pods/components/task-base/component';
-export default TaskComponent.extend();
+export default TaskComponent.extend({
+  init() {
+    this._super(...arguments);
+    this.get('task.paper.decisions').reload();
+  }
+});