+++ engines/tahi_standard_tasks/client/test-support/integration/components/invitation-detail-row-test.js
@@ -2,31 +2,33 @@ import {moduleForComponent, test} from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 import Ember from 'ember';
 
-moduleForComponent('invitation-detail-row',
-                    'Integration | Component | invitation detail row', {
-                      integration: true,
-                      beforeEach: function() {
-                        this.set('update-date', new Date('January 01, 2016'));
-                        this.set('invitation', Ember.Object.create({
-                          declineReason: null,
-                          declined: false,
-                          email: 'jane@example.com',
-                          invitee: Ember.Object.create({
-                            fullName: 'Jane McEdits'
-                          }),
-                          reviewerSuggestions: null,
-                          state: 'pending',
-                          title: 'Awesome Paper!',
-                          updatedAt: this.get('update-date')
-                        }));
-                      }
-                    });
+moduleForComponent('invitation-detail-row', 'Integration | Component | invitation detail row', {
+  integration: true,
+  beforeEach() {
+    this.set('update-date', new Date('January 01, 2016'));
+    this.set('invitation', Ember.Object.create({
+      declineReason: null,
+      declined: false,
+      email: 'jane@example.com',
+      invitee: Ember.Object.create({
+        fullName: 'Jane McEdits'
+      }),
+      reviewerSuggestions: null,
+      state: 'pending',
+      title: 'Awesome Paper!',
+      updatedAt: this.get('update-date')
+    }));
+  }
+});
 
 let template = hbs`{{invitation-detail-row
-                      invitation=invitation}}`;
+                      invitation=invitation
+                      uiState='closed'}}`;
 
 test('displays invitation information if the invite.invited is true', function(assert){
-  this.set('invitation.state', 'invited');
+  this.set('invitation.pending', false);
+  this.set('invitation.invited', true);
+  this.set('invitation.invitedAt', this.get('update-date'));
   this.render(template);
 
   assert.textPresent('.invitation-item-status',
@@ -48,51 +50,59 @@ test('displays invitation email when no invitee present', function(assert){
   assert.textPresent('.invitation-item-full-name', 'jane@example.com');
 });
 
-test('displays delete icon if invite is pending and the row is in the show state',
-  function(assert){
-    this.setProperties({
-      'invitation.pending': true,
-      allowDestroy: true,
-      uiState: 'show'
-    });
-    let openTemplate = hbs`{{invitation-detail-row
-                          allowDestroy=allowDestroy
-                          invitation=invitation
-                          uiState=uiState}}`;
-    this.render(openTemplate);
+test('displays delete icon if invite is pending, the row is in the show state, and in current round', function(assert) {
+  this.setProperties({
+    'invitation.pending': true,
+    currentRound: true,
+    uiState: 'show'
+  });
 
-    assert.elementFound('.invite-remove', 'destroy button is found in show state with AllowDestroy');
+  const openTemplate = hbs`{{invitation-detail-row allowDestroy=allowDestroy
+                                                   invitation=invitation
+                                                   currentRound=currentRound
+                                                   uiState=uiState}}`;
 
-    this.setProperties({
-      'invitation.pending': true,
-      allowDestroy: false,
-      uiState: 'show'
-    });
-    assert.elementNotFound('.invite-remove', 'destroy button not shown if not allowed');
+  this.render(openTemplate);
 
-    this.setProperties({
-      'invitation.pending': false,
-      allowDestroy: true,
-      uiState: 'show'
-    });
-    assert.elementNotFound('.invite-remove', 'destroy button not shown if invitation is not pending');
+  assert.elementFound('.invitation-item-action-delete', 'destroy button is found');
 
-    this.setProperties({
-      'invitation.pending': true,
-      allowDestroy: true,
-      uiState: 'closed'
-    });
-    assert.elementNotFound('.invite-remove', 'destroy button not shown if the row is closed');
-  }
-);
+  this.setProperties({
+    'invitation.pending': true,
+    currentRound: false,
+    uiState: 'show'
+  });
 
+  assert.elementNotFound('.invitation-item-action-delete', 'destroy button not shown if not current round');
+
+  this.setProperties({
+    'invitation.pending': false,
+    currentRound: true,
+    uiState: 'show'
+  });
+
+  assert.elementNotFound('.invitation-item-action-delete', 'destroy button not shown if invitation is not pending');
+
+  this.setProperties({
+    'invitation.pending': true,
+    currentRound: true,
+    uiState: 'closed'
+  });
+
+  assert.elementNotFound('.invitation-item-action-delete', 'destroy button not shown if the row is closed');
+});
 
 test('displays decline feedback when declined', function(assert){
-  this.set('invitation.declined', true);
-  this.set('invitation.declineReason', 'No current availability');
-  this.set('invitation.reviewerSuggestions', 'Jane McReviewer');
-  this.render(template);
+  this.setProperties({
+    'invitation.declined': true,
+    'invitation.declineReason': 'No current availability',
+    'invitation.reviewerSuggestions': 'Jane McReviewer'
+  });
 
-  assert.textPresent('.invitation-decline-reason', 'No current availability');
-  assert.textPresent('.invitation-reviewer-suggestions', 'Jane McReviewer');
+  const openTemplate = hbs`{{invitation-detail-row invitation=invitation
+                                                   uiState='show'}}`;
+
+  this.render(openTemplate);
+
+  assert.textPresent('.invitation-item-decline-info', 'No current availability');
+  assert.textPresent('.invitation-item-decline-info', 'Jane McReviewer');
 });