+++ engines/tahi_standard_tasks/spec/factories/paper_reviewer_task_factory.rb
@@ -1,8 +1,0 @@
-FactoryGirl.define do
-  factory :paper_reviewer_task, class: 'TahiStandardTasks::PaperReviewerTask' do
-    paper
-    phase
-    title 'Invite Reviewers'
-    old_role 'editor'
-  end
-end