+++ engines/tahi_standard_tasks/spec/features/invite_reviewer_spec.rb
@@ -15,17 +15,14 @@ feature "Invite Reviewer", js: true do
 
   before do
     assign_journal_role journal, editor, :editor
-    paper.paper_roles.create user: editor, old_role: PaperRole::COLLABORATOR
-    task.add_participant(editor)
-
     login_as(editor, scope: :user)
     visit "/"
   end
 
   scenario "Editor can invite any user as a reviewer to a paper" do
     overlay = Page.view_task_overlay(paper, task)
-    overlay.paper_reviewers = [reviewer1]
-    expect(overlay).to have_reviewers reviewer1.full_name
+    overlay.invited_users = [reviewer1]
+    expect(overlay).to have_invitees reviewer1.full_name
 
     # Already invited users don't show up again the search
     overlay.fill_in 'invitation-recipient', with: 'Hen'
@@ -38,7 +35,7 @@ feature "Invite Reviewer", js: true do
 
   scenario "displays invitations from the latest round of revisions" do
     overlay = Page.view_task_overlay(paper, task)
-    overlay.paper_reviewers = [reviewer1]
+    overlay.invited_users = [reviewer1]
     expect(overlay.active_invitations_count(1)).to be true
 
     register_paper_decision(paper, 'minor_revision')
@@ -46,9 +43,78 @@ feature "Invite Reviewer", js: true do
 
     overlay.reload
     overlay = Page.view_task_overlay(paper, task)
-    overlay.paper_reviewers = [reviewer3, reviewer2]
+    overlay.invited_users = [reviewer3, reviewer2]
     expect(overlay.expired_invitations_count(1)).to be true
     expect(overlay.active_invitations_count(2)).to be true
     expect(overlay.total_invitations_count(3)).to be true
+  end
+
+  scenario "links alternate candidates with other potential reviewers" do
+    overlay = Page.view_task_overlay(paper, task)
+    overlay.invited_users = [reviewer1]
+    expect(overlay).to have_invitees reviewer1.full_name
+
+    overlay.fill_in 'invitation-recipient', with: reviewer2.email
+    overlay.find('.invitation-email-entry-button').click
+
+    overlay.select_first_alternate
+    find('.invitation-save-button').click
+    expect(page.find('.alternate-link-icon')).to be_present
+  end
+
+  scenario "prevents invitations while a review is invited in the subqueue" do
+    overlay = Page.view_task_overlay(paper, task)
+    overlay.invited_users = [reviewer1]
+    overlay.add_to_queue(reviewer2)
+    overlay.find('.invitation-item-action-edit').click
+    overlay.select_first_alternate
+    overlay.find('.invitation-save-button').click
+
+    header = find('.invitation-item-header', text: reviewer2.first_name)
+    expect(header).to have_css('.invitation-item-action-send.invitation-item-action--disabled')
+  end
+
+  scenario "does not disable other invitations in the main queue when another invitation is invited/accepted" do
+    overlay = Page.view_task_overlay(paper, task)
+    # invited user
+    overlay.invited_users = [reviewer1]
+    # pending user
+    overlay.add_to_queue(reviewer2)
+    overlay.find('.invitation-item-action-edit').click
+    overlay.find('.invitation-save-button').click
+
+    header = find('.invitation-item-header', text: reviewer2.first_name)
+    expect(header).to have_no_css('.invitation-item-action-send.invitation-item-action--disabled')
+  end
+
+  scenario "edits an invitation" do
+    overlay = Page.view_task_overlay(paper, task)
+    overlay.add_to_queue(reviewer1)
+
+    # Life is not lost by dying; life is lost minute by minute,
+    # day by dragging day, in all the thousand small uncaring ways.
+    overlay.find('.invitation-item-header').click
+    overlay.find('.invitation-item-header').click
+
+    overlay.find('.invitation-item-action-edit').click
+    overlay.invitation_body = 'New body'
+    overlay.find('.invitation-save-button').click
+    expect(overlay.find('.invitation-show-body')).to have_text('New body')
+  end
+
+  scenario "deletes only a pending invitation" do
+    overlay = Page.view_task_overlay(paper, task)
+
+    # invited
+    overlay.invited_users = [reviewer1]
+    find('.invitation-item-header', text: reviewer1.first_name).click
+    expect(overlay).to have_no_css('.invitation-item-action-delete')
+
+    # pending
+    overlay.add_to_queue(reviewer2)
+    header = find('.invitation-item-header', text: reviewer2.first_name)
+    expect(header).to have_css('.invitation-item-action-delete')
+    header.find('.invitation-item-action-delete').click
+    expect(overlay).to have_no_css('.invitation-item-header', text: reviewer2.first_name)
   end
 end