+++ engines/tahi_standard_tasks/spec/support/pages/overlays/invite_editor_overlay.rb
@@ -1,0 +1,3 @@
+# Shows up on the PaperEditorTask
+class InviteEditorOverlay < BaseInviteUsersOverlay
+end